added option menu for every key
you can change the id, effort and keypress values on every layer. just click on any key to open the menu, select the layer you want to edit and edit it in the text entry below. change as many values as you like and then press ok to save or cancel to discard the changes